- All datasets Epoch/Labels are on type float32  On branch dev  Your branch is up to date with 'origin/dev'.
 Changes to be committed:
	modified:   aawedha/io/bci_comp_iv_2a.py
	modified:   aawedha/io/beta.py
	modified:   aawedha/io/epfl.py
	modified:   aawedha/io/essex.py
	modified:   aawedha/io/exoskeleton.py
	modified:   aawedha/io/freiburg_online.py
	modified:   aawedha/io/multiday.py
	modified:   aawedha/io/openbmi.py
	modified:   aawedha/io/perturbations.py
	modified:   aawedha/io/san_diego.py
	modified:   aawedha/io/tsinghua.py

+++ aawedha/io/bci_comp_iv_2a.py
@@ -60,8 +60,8 @@ class Comp_IV_2a(DataSet):
             X.append(x)
             Y.append(y)
         
-        X = np.array(X)
-        Y = np.array(Y)
+        X = np.array(X, dtype=np.float32)
+        Y = np.array(Y, dtype=np.float32)
         return X, Y
     
     def load_raw_legacy(self, path=None, mode='',
