+++ aawedha/io/beta.py
@@ -184,8 +184,8 @@ class Beta(DataSet):
             del eeg
             del y
 
-        X = np.array(X)
-        Y = np.array(Y).squeeze()
+        X = np.array(X, dtype=np.float32)
+        Y = np.array(Y, dtype=np.float32).squeeze()
         return X, Y, subj_info  
 
     def download_raw(self, store_path=None):
