+++ aawedha/io/freiburg_online.py
@@ -96,8 +96,8 @@ class FreiburgOnline(DataSet):
 
         samples, channels, trials = X[0].shape
         X = np.array(X).reshape(
-            (n_subjects, samples, channels, trials), order='F')
-        Y = np.array(Y).reshape((n_subjects, trials), order='F')
+            (n_subjects, samples, channels, trials), order='F', dtype=np.float32)
+        Y = np.array(Y).reshape((n_subjects, trials), order='F', dtype=np.float32) 
 
         return X, Y
