+++ aawedha/io/multiday.py
@@ -237,8 +237,8 @@ class MultiDay(DataSet):
             X.append(np.concatenate(x_subj, axis=-1))
             Y.append(np.concatenate(y_subj, axis=-1))
 
-        X = np.array(X)
-        Y = np.array(Y).squeeze()
+        X = np.array(X, dtype=np.float32)
+        Y = np.array(Y, dtype=np.float32).squeeze()
         return X, Y
 
     def download_raw(self, store_path):
